+++ akka-loader/src/main/scala/chpp/players/PlayersRequest.scala
@@ -5,7 +5,7 @@ import chpp.{AbstractRequest, OauthTokens, RequestCreator}
 
 case class PlayersRequest(actionType: Option[String] = None,
                           orderBy: Option[String] = None,
-                          teamId: Option[Int] = None,
+                          teamId: Option[Long] = None,
                           includeMatchInfo: Option[Boolean] = None) extends AbstractRequest {
   override def createRequest()(implicit oauthTokens: OauthTokens): HttpRequest = {
     val map = RequestCreator.params("players", "2.4",
