+++ akka-loader/src/main/scala/chpp/teamdetails/TeamDetailsRequest.scala
@@ -3,7 +3,7 @@ package chpp.teamdetails
 import akka.http.scaladsl.model.HttpRequest
 import chpp.{AbstractRequest, OauthTokens, RequestCreator}
 
-case class TeamDetailsRequest(teamId: Option[Int] = None,
+case class TeamDetailsRequest(teamId: Option[Long] = None,
                               userId: Option[Int] = None,
                               includeDomesticFlags: Option[Boolean] = None,
                               includeFlags: Option[Boolean] = None,
